DevOps/gantt_api/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
PyJWT==1.7.1
httpx==0.23.1
